deploy/pypi_requirement.txt
mysql-python==1.2.5
DBUtils==1.1
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-Login==0.3.2
Flask-HTTPAuth (3.2.2)

pyDes==2.0.1
rratelimit==0.0.4
jsonschema==2.4.0
supervisor==3.1.3
unirest==1.1.7
ansible==2.4.6
python-ldap==2.4.28
libvirt-python==1.2.7
requests==2.20.1
pykafka==2.5.0
kazoo==2.2.1
IPy==0.83
netaddr==0.7.18



amqp (1.4.9)
anyjson (0.3.3)
billiard (3.3.0.23)
celery (3.1.23)
cffi (1.9.1)
click (6.7)
cryptography (1.7.2)
enum34 (1.1.6)
greenlet (0.4.12)
gevent (1.2.1)
gunicorn (18.0)
idna (2.5)
ipaddress (1.0.18)
IPy (0.83)
itsdangerous (0.24)
Jinja2 (2.8.1)
jsonschema (2.4.0)
kazoo (2.2.1)
kombu (3.0.35)
libxml2-python (2.6.15)
lockfile (0.12.2)
MarkupSafe (0.23)
meld3 (1.0.2)
netaddr (0.7.19)
paramiko (2.1.2)
pip (9.0.1)
poster (0.8.1)
pyasn1 (0.2.3)
pycparser (2.17)
pycrypto (2.6.1)
pyDes (2.0.1)
python-engineio (1.3.2)
pytz (2016.6.1)
PyYAML (3.12)
requests (2.20.1)
rratelimit (0.0.4)
setuptools (26.1.1)
six (1.10.0)
SQLAlchemy (1.0.15)
supervisor (3.1.3)
tabulate (0.7.7)
Unirest (1.1.7)
virtualenv (15.1.0)
websockify (0.8.0)
Werkzeug (0.11.15)
wheel (0.29.0)
